// nextjs/ucenter-web-ssr/src/api/ucenter/types.gen.ts
export type UserResponse = {
    /**
     * 头像
     */
    avatar?: string;
    /**
     * 头像是否审核中 ture=审核中，false=审核通过
     */
    avatarPending?: boolean;
    /**
     * 余额计价币种
     */
    balanceCurrency?: string;
    /**
     * 手机区号
     */
    countryCode?: string;
    /**
     * 创建时间（注册时间）
     */
    createdAt?: string;
    /**
     * 币种
     */
    currency?: string;
    /**
     * 用户设备
     */
    device?: string;
    /**
     * 租户ID
     */
    domainId?: string;
    /**
     * 邮箱
     */
    email?: string;
    emailValidate?: boolean;
    existLoginPsd?: boolean;
    /**
     * vip等级 0- 3
     */
    honorLevel?: number;
    /**
     * 托管子账号绑定交易团队用户ID
     */
    hostedUserId?: string;
    /**
     * 用户ID
     */
    id?: string;
    /**
     * 邮箱是否已激活
     */
    isEmailValidate?: boolean;
    /**
     * 充币是否需要验证
     */
    isNeedDepositValidate?: boolean;
    /**
     * 手机是否已激活
     */
    isPhoneValidate?: boolean;
    /**
     * 语言
     */
    language?: string;
    /**
     * 上次登录时间
     */
    lastLoginAt?: string;
    /**
     * 上次登录国家
     */
    lastLoginCountry?: string;
    needDepositValidate?: boolean;
    /**
     * 昵称
     */
    nickname?: string;
    /**
     * 昵称是否审核中 ture=审核中，false=审核通过
     */
    nicknamePending?: boolean;
    /**
     * 1:oes已绑定；0：oes待绑定；其它值(包括-1)：无法绑定oes；默认值-1
     */
    oesBound?: number;
    /**
     * true:允许创建oes子账号；false/null：不允许创建oes子账号
     */
    oesSubEnabled?: boolean;
    /**
     * parentId
     */
    parentId?: string;
    /**
     * 开通的交易类型
     */
    permissionTrades?: Array<string>;
    /**
     * 手机号
     */
    phone?: string;
    phoneValidate?: boolean;
    /**
     * 推荐码
     */
    referralCode?: string;
    /**
     * 用户站点类型global主站，turkey土耳其站
     */
    siteType?: string;
    /**
     * 用户状态 1（未激活状态）2（新用户，已激活状态）3 （冻结状态）-1（已删除）0（禁用）9(迁移完成状态) 10 (升级交易密码完成状态，老用户激活状态)
     */
    status?: number;
    /**
     * subAccount
     */
    subAccount?: string;
    /**
     * 当前会员级别 1 - 3
     */
    subLevel?: number;
    /**
     * 子账号类型
     */
    subType?: number;
    /**
     * broker租户ID
     */
    tenant?: string;
    /**
     * 时区
     */
    timeZone?: string;
    /**
     * 开通的交易类型
     */
    tradeType?: number;
    /**
     * 用户类型 1（普通用户）2 （内部账号）3（子账号）4 (测试账号) 5 (返佣做市商) 6 (项目方) 7(项目方做市商) 8 (固定费率做市商)
     */
    type?: number;
    /**
     * 用户UID, 展示给用户
     */
    uid?: number;
    /**
     * web端用户登录设备限制数 -1为不限制
     */
    webDeviceLimit?: number;
};
// ...
export type SetUserLocaleUsingPostData = {
    body?: never;
    headers: {
        /**
         * X-DOMAIN-ID
         */
        'X-DOMAIN-ID': string;
    };
    path?: never;
    query?: {
        /**
         * 余额计价币种
         */
        balanceCurrency?: string;
        /**
         * 汇率
         */
        currency?: string;
        /**
         * 语言
         */
        language?: string;
        /**
         * 时区
         */
        timeZone?: string;
    };
    url: '/user/locale';
};