// nextjs/ucenter-web-ssr/src/api/growth-ucenter/types.gen.ts
/**
      * This file is auto-generated by @kucoin/api-generator
      * Owner: sean.shi@kupotech.com
      */

export type CampaignI18nInfoVo = {
    /**
     * 活动描述
     */
    desc?: string;
    /**
     * 活动id
     */
    id?: string;
    /**
     * 活动跳转链接
     */
    jumpLink?: string;
    /**
     * 活动海报链接
     */
    posterUrl?: string;
    /**
     * 活动副标题
     */
    subTitle?: string;
    /**
     * 活动标题
     */
    title?: string;
};

export type GenericResult_InvitationUserResponse_ = {
    code?: string;
    data?: InvitationUserResponse;
    msg?: string;
    retry?: boolean;
    success?: boolean;
};

export type InvitationUserResponse = {
    /**
     * 头像
     */
    avatar?: string;
    /**
     * 活动福利
     */
    campaigns?: Array<CampaignI18nInfoVo>;
    /**
     * 被邀请者返现比例
     */
    cashbackRatio?: number;
    /**
     * 返现比例展示文案
     */
    cashbackRatioText?: string;
    /**
     * 邮箱
     */
    email?: string;
    /**
     * 邀请人渠道：TobReferral、TocReferral
     */
    inviterCategory?: string;
    /**
     * 邀请人uid
     */
    inviterUid?: number;
    /**
     * 邀请人名称（昵称/邮箱/手机号）
     */
    inviterUser?: string;
    /**
     * 邀请人id
     */
    inviterUserId?: string;
    /**
     * 语言
     */
    lang?: string;
    /**
     * 邀请寄语
     */
    message?: string;
    /**
     * 是否多级合伙人
     */
    multi?: boolean;
    /**
     * 是否多级合伙人根合伙人
     */
    multiRoot?: boolean;
    /**
     * 昵称
     */
    nickname?: string;
    /**
     * 手机号码
     */
    phone?: string;
    /**
     * 邀请码
     */
    rcode?: string;
    /**
     * 邀请码类型
     */
    rcodeType?: 'broker' | 'tob' | 'toc';
};

export type GetUserByRcodeUsingGetData = {
    body?: never;
    headers: {
        /**
         * X-USER-ID
         */
        'X-USER-ID'?: string;
        /**
         * X-KC-CDN-SITE-TYPE
         */
        'X-KC-CDN-SITE-TYPE': string;
        /**
         * LANG
         */
        LANG: string;
        /**
         * X-IP
         */
        'X-IP': string;
        /**
         * X-LOCATION-CODE
         */
        'X-LOCATION-CODE'?: string;
    };
    path?: never;
    query: {
        /**
         * rcode
         */
        rcode: string;
    };
    url: '/invitation/user-by-rcode';
};

export type GetUserByRcodeUsingGetResponses = {
    /**
     * OK
     */
    200: GenericResult_InvitationUserResponse_;
};

export type GetUserByRcodeUsingGetResponse = GetUserByRcodeUsingGetResponses[keyof GetUserByRcodeUsingGetResponses];

export type ClientOptions = { baseURL: string; };